+++ forge/src/main/java/io/github/prismwork/emitrades/forge/EMITradesForge.java
@@ -1,8 +1,0 @@
-package io.github.prismwork.emitrades.forge;
-
-import net.minecraftforge.fml.common.Mod;
-
-@Mod("emitrades")
-public final class EMITradesForge {
-    public EMITradesForge() {}
-}