+++ neoforge/src/main/java/io/github/prismwork/emitrades/neoforge/EMITradesNeoForge.java
@@ -1,0 +1,10 @@
+package io.github.prismwork.emitrades.neoforge;
+
+
+import net.neoforged.fml.common.Mod;
+
+@SuppressWarnings("unused")
+@Mod("emitrades")
+public final class EMITradesNeoForge {
+    public EMITradesNeoForge() {}
+}