+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-* Russian translation by Disketaa
+* Updated to 1.20.4
+* Dropped Forge support, added (specific) NeoForge support
+* Some changes to the tipped arrow & enchanted thingy trades
+  * Now the output slot shows the possible items as if it is a `ListEmiIngredient` instead of showing simply "random effects" or annoyingly listing the possible enchantments in text